select_info had the wrong size (uint8 instead of uint16). Corrected POLLNVAL handling.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@1750 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kernel/core/fs/vfs_select.h
@@ -6,7 +6,7 @@
 #define VFS_SELECT_H
 
 typedef struct select_info {
-	uint8	events;
+	uint16	events;
 } select_info;
 
 typedef struct select_sync {
